Renomeado ProdutoController.java para ProdutoRestAdapter.java Adicionado o metodo create
